+++ boxsdk/auth/jwt_auth.py
@@ -8,12 +8,13 @@ import string
 
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
+from cryptography.hazmat.primitives.asymmetric.rsa import RSAPrivateKey
 import jwt
-from six import string_types, text_type
+from six import binary_type, string_types, raise_from, text_type
 
 from .oauth2 import OAuth2
 from ..object.user import User
-from ..util.compat import total_seconds
+from ..util.compat import NoneType, total_seconds
 
 
 class JWTAuth(OAuth2):
@@ -28,7 +29,7 @@ class JWTAuth(OAuth2):
             client_secret,
             enterprise_id,
             jwt_key_id,
-            rsa_private_key_file_sys_path,
+            rsa_private_key_file_sys_path=None,
             rsa_private_key_passphrase=None,
             user=None,
             store_tokens=None,
@@ -37,8 +38,12 @@ class JWTAuth(OAuth2):
             access_token=None,
             network_layer=None,
             jwt_algorithm='RS256',
+            rsa_private_key_data=None,
     ):
         """Extends baseclass method.
+
+        Must pass exactly one of either `rsa_private_key_file_sys_path` or
+        `rsa_private_key_data`.
 
         If both `enterprise_id` and `user` are non-`None`, the `user` takes
         precedence when `refresh()` is called. This can be overruled with a
@@ -68,13 +73,13 @@ class JWTAuth(OAuth2):
         :type jwt_key_id:
             `unicode`
         :param rsa_private_key_file_sys_path:
-            Path to an RSA private key file, used for signing the JWT assertion.
+            (optional) Path to an RSA private key file, used for signing the JWT assertion.
         :type rsa_private_key_file_sys_path:
             `unicode`
         :param rsa_private_key_passphrase:
             Passphrase used to unlock the private key. Do not pass a unicode string - this must be bytes.
         :type rsa_private_key_passphrase:
-            `str` or None
+            `bytes` or None
         :param user:
             (optional) The user to authenticate, expressed as a Box User ID or
             as a :class:`User` instance.
@@ -120,8 +125,20 @@ class JWTAuth(OAuth2):
             Which algorithm to use for signing the JWT assertion. Must be one of 'RS256', 'RS384', 'RS512'.
         :type jwt_algorithm:
             `unicode`
+        :param rsa_private_key_data:
+            (optional) Contents of RSA private key, used for signing the JWT assertion. Do not pass a
+            unicode string. Can pass a byte string, or a file-like object that returns bytes, or an
+            already-loaded `RSAPrivateKey` object.
+        :type rsa_private_key_data:   `bytes` or :class:`io.IOBase` or :class:`RSAPrivateKey`
         """
         user_id = self._normalize_user_id(user)
+        rsa_private_key = self._normalize_rsa_private_key(
+            file_sys_path=rsa_private_key_file_sys_path,
+            data=rsa_private_key_data,
+            passphrase=rsa_private_key_passphrase,
+        )
+        del rsa_private_key_data
+        del rsa_private_key_file_sys_path
         super(JWTAuth, self).__init__(
             client_id,
             client_secret,
@@ -132,12 +149,7 @@ class JWTAuth(OAuth2):
             refresh_token=None,
             network_layer=network_layer,
         )
-        with open(rsa_private_key_file_sys_path, 'rb') as key_file:
-            self._rsa_private_key = serialization.load_pem_private_key(
-                key_file.read(),
-                password=rsa_private_key_passphrase,
-                backend=default_backend(),
-            )
+        self._rsa_private_key = rsa_private_key
         self._enterprise_id = enterprise_id
         self._jwt_algorithm = jwt_algorithm
         self._jwt_key_id = jwt_key_id
@@ -295,3 +307,54 @@ class JWTAuth(OAuth2):
         else:
             new_access_token = self.authenticate_user()
         return new_access_token, None
+
+    @classmethod
+    def _normalize_rsa_private_key(cls, file_sys_path, data, passphrase=None):
+        if len(list(filter(None, [file_sys_path, data]))) != 1:
+            raise TypeError("must pass exactly one of either rsa_private_key_file_sys_path or rsa_private_key_data")
+        if file_sys_path:
+            with open(file_sys_path, 'rb') as key_file:
+                data = key_file.read()
+        if hasattr(data, 'read') and callable(data.read):
+            data = data.read()
+        if isinstance(data, text_type):
+            try:
+                data = data.encode('ascii')
+            except UnicodeError:
+                raise_from(
+                    TypeError("rsa_private_key_data must contain binary data (bytes/str), not a text/unicode string"),
+                    None,
+                )
+        if isinstance(data, binary_type):
+            passphrase = cls._normalize_rsa_private_key_passphrase(passphrase)
+            return serialization.load_pem_private_key(
+                data,
+                password=passphrase,
+                backend=default_backend(),
+            )
+        if isinstance(data, RSAPrivateKey):
+            return data
+        raise TypeError(
+            'rsa_private_key_data must be binary data (bytes/str), '
+            'a file-like object with a read() method, '
+            'or an instance of RSAPrivateKey, '
+            'but got {!r}'
+            .format(data.__class__.__name__)
+        )
+
+    @staticmethod
+    def _normalize_rsa_private_key_passphrase(passphrase):
+        if isinstance(passphrase, text_type):
+            try:
+                return passphrase.encode('ascii')
+            except UnicodeError:
+                raise_from(
+                    TypeError("rsa_private_key_passphrase must contain binary data (bytes/str), not a text/unicode string"),
+                    None,
+                )
+        if not isinstance(passphrase, (binary_type, NoneType)):
+            raise TypeError(
+                "rsa_private_key_passphrase must contain binary data (bytes/str), got {!r}"
+                .format(passphrase.__class__.__name__)
+            )
+        return passphrase